Commit Matheus

+++ app/Controllers/Admin/Usuarios.php
@@ -18,7 +18,8 @@ class Usuarios extends BaseController
 
         $data =[
             'titulo' => 'Listando os usuarios',
-            'usuarios' => $this->usuarioModel->withDeleted(true)->findAll()
+            'usuarios' => $this->usuarioModel->withDeleted(true)->paginate(10),
+            'pager' => $this->usuarioModel->pager,
 
         ];
 
@@ -62,7 +63,7 @@ class Usuarios extends BaseController
 
     public function cadastrar (){
 
-        if($this->request->getMethod() === 'post'){
+        if($this->request->getPost()){
 
             $usuario = new Usuario($this->request->getPost());
 
@@ -103,6 +104,10 @@ class Usuarios extends BaseController
     public function editar($id=null){
 
         $usuario = $this->buscaUsuarioOu404($id);
+
+        if($usuario->deletado_em != null){
+            return redirect()->back()->with('info',"O usuário $usuario->nome encontra-se excluído. Portanto, não é possível editá-lo.");
+        }
        
         $data= [
 
@@ -118,6 +123,10 @@ class Usuarios extends BaseController
         if($this->request->getMethod() === 'post'){
 
             $usuario=$this->buscaUsuarioOu404($id);
+
+            if($usuario->deletado_em != null){
+                return redirect()->back()->with('info',"O usuário $usuario->nome encontra-se excluído. Portanto, não é possível editá-lo.");
+            }
 
             $post= $this->request->getPost();
             
@@ -158,7 +167,14 @@ class Usuarios extends BaseController
 
         $usuario = $this->buscaUsuarioOu404($id);
 
-        if($this->request->getMethod() === 'post'){
+        if($usuario->deletado_em != null){
+            return redirect()->back()->with('info',"O usuário $usuario->nome encontra-se ja excluído.");
+        }
+
+        if($usuario->is_admin){
+            return redirect()->back()->with('info','Não é possivel excluirum usuário <b>Administrador</b>');
+        }
+        if ($this->request->getMethod() === 'post') {
 
             $this->usuarioModel->delete($id);
             return redirect()->to(site_url('admin/usuarios'))->with('sucesso',"Usuario $usuario->nome excluido com Sucesso!");
@@ -173,8 +189,29 @@ class Usuarios extends BaseController
     }
 
 
+    public function desfazerExclusao($id=null){
+
+        $usuario = $this->buscaUsuarioOu404($id);
+        
+        if($usuario->deletado_em == null){
+            return redirect()->back()->with('info','Apenas usuários excluidos podem ser recuperado.');
+        }
+
+        if($this->usuarioModel->desfazerExclusao($id)){
+            return redirect()->back()->with('sucesso','Exclusão desfeita com Sucesso!');
+        }else{
+            return redirect()->back()
+                    ->with('errors_model', $this->usuarioModel->errors())
+                    ->with('atencao','Por favor verifique os erros abaixo')
+                    ->withInput();
+        }
+
+    }
+
+
+
     private function buscaUsuarioOu404(int $id=null){
-        if(!$id || !$usuario=$this->usuarioModel->where('id',$id)->first()){
+        if(!$id || !$usuario=$this->usuarioModel->withDeleted(true)->where('id',$id)->first()){
             throw \CodeIgniter\Exceptions\PageNotFoundException::forPageNotFound("Não encontramos o usuario $id");
         }
         return $usuario;
